Cook Oh Salmon
INGREDIENTS:

1  (1 1/2-pound) skin-on or skinless salmon fillet
12  fresh cilantro sprigs
2  tablespoons pure maple syrup
2  tablespoons whole-grain Dijon mustard
1  tablespoon mayonnaise
 Kosher salt and freshly ground black pepper
 Lemon wedges, for serving

PREPARATION

Remove salmon from the refrigerator. Heat oven to 325 degrees.
Bundle the cilantro sprigs by their stems and hold them tightly, then slice the stems crosswise until you get to the leaves. Reserve leaves for garnish. Transfer sliced stems to a small bowl and stir in the maple syrup, mustard and mayonnaise until well mixed.
Season the salmon all over with salt and pepper and place in a baking dish, skin-side down if there is skin. Slather the maple sauce all over the top.
Bake until a paring knife slides into the center with only a little resistance, 15 to 20 minutes. When you remove the knife and touch the blade to your upper lip, it should feel very warm but not hot. The salmon will continue to heat through out of the oven while in the baking dish. Top with the reserved cilantro leaves, and squeeze lemon wedges all over just before serving.

https://cooking.nytimes.com/guides/19-how-to-cook-salmon